Only run CI steps on macOS runners (but also future-proof somewhat)
The work to enable Linux casks has resulted in the build failing because
macOS-only casks with different SHAs for Intel and ARM end up evaluating to
`nil` for Linux if the checks are run on a Linux machine.

This borrows from the homebrew-cask repo's build steps and dynamically generates
a build matrix--if there are no Linux casks, it won't build on Linux.

If I read it right, though, adding a single Linux cask will trigger Linux builds
for all casks from the main branch, so this isn't a permanent solution.

Deciding to add the increased complexity because, while a `depends_on :macos` is
implied right now, if that ever changes, this repo will start building on Linux
again and I'll see the build failures. Anyway...